added check for public folder existence

+++ __hugo_web/publish.sh
@@ -6,7 +6,9 @@ cd $HUGO_DIR
 hugo -d hugo_tmp --theme="liquorice"
 mv hugo_tmp/* ../
 rm -rf hugo_tmp
+if [ -d public ]; then
 rm -rf public
+fi
 cd ..
 git add -A .
